mise a jour des routes, protection des routes avec les middlewares pour la sécurité

+++ blog/routes/web.php
@@ -1,7 +1,10 @@
 <?php
 
 use App\Http\Controllers\ArticlesController;
+use App\Http\Controllers\Auth\RegisterController;
+use App\Http\Controllers\Auth\SessionsController;
 use App\Http\Controllers\PagesController;
+use App\Http\Controllers\UserController;
 use Illuminate\Support\Facades\Route;
 
 Route::get('/', [PagesController::class, 'index']);
@@ -9,9 +12,20 @@ Route::get('/contact-us', [PagesController::class, 'contact']);
 Route::get('/about-us', [PagesController::class, 'about']);
 Route::get('/articles', [ArticlesController::class, 'index']);
 Route::get('/show-article/{id}', [ArticlesController::class, 'show']);
-Route::get('/articles/create', [ArticlesController::class, 'create']);
-Route::post('/articles/create', [ArticlesController::class, 'store']);
-Route::get('article/{article}/edit', [ArticlesController::class, 'edit']);
-Route::patch('/article/{article}/edit', [ArticlesController::class, 'update']);
-Route::delete('article/{article}/delete', [ArticlesController::class, 'delete']);
-// Route::get('show-article/{article}', [ArticlesController::class, 'show']);
+
+Route::get('/articles/create', [ArticlesController::class, 'create'])->middleware('auth');
+Route::post('/articles/create', [ArticlesController::class, 'store'])->middleware('auth');
+
+Route::get('/article/{article}/edit', [ArticlesController::class, 'edit'])->middleware('auth');
+Route::patch('/article/{article}/edit', [ArticlesController::class, 'update'])->middleware('auth');
+Route::delete('article/{article}/delete', [ArticlesController::class, 'delete'])->middleware('auth');
+
+// Auth
+Route::get('/register', [RegisterController::class, 'index'])->name('register')->middleware('guest');
+Route::post('/register', [RegisterController::class, 'create'])->middleware('guest');
+Route::get('/login', [SessionsController::class, 'index'])->name('login')->middleware('guest');
+Route::post('/login', [SessionsController::class, 'authenticate'])->middleware('guest');
+Route::post('/logout', [SessionsController::class, 'logout'])->name('logout')->middleware('auth');
+
+// profile
+Route::get('/profile', [UserController::class, 'index'])->name('profile')->middleware('auth');
